--- src/plugins/three/ImageFormatPlugin.d.ts
@@ -1,19 +1,0 @@
-export class DeepZoomImagePlugin {
-
-	constructor( options: {
-		center?: boolean,
-		pixelSize?: number,
-	} );
-
-}
-
-export class XYZTilesPlugin {
-
-	constructor( options: {
-		center?: boolean,
-		pixelSize?: number,
-		levels?: number,
-		tileDimension?: number,
-	} );
-
-}
