+++ src/plugins/three/images/EllipsoidProjectionTilesPlugin.d.ts
@@ -1,0 +1,22 @@
+
+export class XYZTilesPlugin {
+
+	constructor( options: {
+		center?: boolean,
+		pixelSize?: number,
+		levels?: number,
+		tileDimension?: number,
+		projection?: 'ellipsoid' | 'planar',
+	} );
+
+}
+
+export class TMSTilesPlugin {
+
+	constructor( options: {
+		center?: boolean,
+		pixelSize?: number,
+		projection?: 'ellipsoid' | 'planar',
+	} );
+
+}
