+++ src/plugins/three/images/ImageFormatPlugin.d.ts
@@ -1,0 +1,8 @@
+export class DeepZoomImagePlugin {
+
+	constructor( options: {
+		center?: boolean,
+		pixelSize?: number,
+	} );
+
+}
